+++ src/pages/Login.js
@@ -1,9 +1,13 @@
 import React from "react";
 import { useForm } from "react-hook-form";
-import { TextField, Button, Box, Typography, Link } from "@mui/material";
+import { Box, Typography, Link } from "@mui/material";
 import { signInWithEmailAndPassword } from "firebase/auth";
 import { auth } from "../firebase";
 import { Link as RouterLink, useNavigate } from "react-router-dom";
+
+import Password from "../components/Pssword";
+import Email from "../components/Email";
+import SubmitButton from "../components/SubmitButton";
 
 const Login = () => {
   const navigate = useNavigate();
@@ -31,29 +35,9 @@ const Login = () => {
         ログイン
       </Typography>
       <form onSubmit={handleSubmit(onSubmit)}>
-        <Box sx={{ mb: 2 }}>
-          <TextField
-            {...register("email", { required: "メールアドレスは必須です" })}
-            label="メールアドレス"
-            placeholder="Email"
-            fullWidth
-            error={Boolean(errors.email)}
-            helperText={errors.email?.message}
-          />
-        </Box>
-        <Box sx={{ mb: 2 }}>
-          <TextField
-            {...register("password", { required: "パスワードは必須です" })}
-            type="password"
-            label="パスワード"
-            fullWidth
-            error={Boolean(errors.password)}
-            helperText={errors.password?.message}
-          />
-        </Box>
-        <Button type="submit" variant="contained" fullWidth>
-          ログイン
-        </Button>
+        <Email errors={errors} register={register} />
+        <Password errors={errors} register={register} />
+        <SubmitButton title="ログイン" />
       </form>
       <Typography variant="body2" align="center" sx={{ mt: 2 }}>
         アカウントをお持ちでないですか？{" "}
